Finished testing Law1.

+++ src/ACEtk/interpretation/ContinuousEnergyNeutron/EnergyDistributionLaws/TabularEquiprobableEnergyBins/src/ctor.hpp
@@ -1,16 +1,16 @@
 TabularEquiprobableEnergyBins( ENDFInterpolationParameters&& interp,
                                Table::Slice energies,
-                               int NET, Table::Slice outgoingEnergies):
+                               size_t NET, Table::Slice outgoingEnergies):
   interpolationParameters_( interp ),
   incidentEnergies_( energies ),
   outgoingEnergyTables_( make( outgoingEnergies, incidentEnergies_.size(),
                               NET ) )
-    // outgoingEnergies | ranges::view::chunk( incidentEnergies_.size() ) )
-{ }
+{ 
+}
 
 template< typename Range >
 TabularEquiprobableEnergyBins( Range& NBT, Range& INT, Range& energies, 
-                               int NET, Range& outgoingEnergies)
+                               size_t NET, Range& outgoingEnergies)
   try:
     TabularEquiprobableEnergyBins( ENDFInterpolationParameters( NBT, INT ),
         Table::Slice{ energies.begin(), energies.end() },
